Introduce DIAMETER_rfc5447_Templates.ttcn
Add templates for RFC 5447 MIP6-Feature-Vector.
This AVP is a bitmask which as only a few bits defined in RFC 5447, with
other interfaces using this AVP adding interface-specific bits in the
spec of each interface.

The templates are added in a new separate file with the aim of start
splitting the tons of AVPs originating from different specs in order to
be able to quickly identify them and avoid confusion.

Change-Id: I0fc646e5354d78283a2f3e1b9bb9c4688cf744a1