hnbgw: Introduce mp_hnbgw_timer_x31
The default value of the timer changed recently, and made some tests
which were using some timer hardcoded values fail.
Add this configurable option to adapt the test behavior to the
osmo-hnbgw configuration. This way we can also set it to lower values in
order to avoid waiting too much time.

Change-Id: I176ef96e193f2ca39077bcee3a2187768ddb45ce