RSL_Emulation: allow RSL_DchanHdlr sandbox escape for both roles
This feature was initially added in 1470fc668 for the RSL_Emulation_CT
operating in the BTS role.  In a subsequent patch I also need to escape
the sandbox, but in my case the RSL_Emulation_CT is playing the BSC role.

Change-Id: I42557f85d3e5e7ef7147a87293f5dd1716884534
Related: 1470fc668 "allow ASP_RSL_UD on MSC_ConnHdlr RSL"
Related: SYS#6655