PCUIF_Types: make PCUIF_info_ind field bts_model optional
The field bts_model has been introduced with PCUIF v12, which is now
used in master. In latest we still use PCUIF v11, where the bts_model
field is not present. This means that when we test against osmo-bts
latest there will be an error parsing PCUIF_info_ind. A solution to this
problem is to declare the bts_model field as an optional field until
PCUIF v12 is also used in latest.

Related: OS#6191
Change-Id: I45947f3bddaccafee82824ebb1b8494eabc67e7a