PCU_Tests: tell PCU a more realistic BTS model
In ts_PCUIF_INFO_default we set the bts_model member to
PCU_IF_BTS_MODEL_UNSPEC. This means that the PCU will not know the BTS
model and thus it is not able to comply to the specific pecularities of
the BTS model that is used. In PCU_Tests.ttcn we simulate the behavior
of an OsmoBTS with OsmoTRX, therefore we should set the BTS model to
PCU_IF_BTS_MODEL_TRX.

This will also fix the problems we have with unexpected DUMMY messages
we currently receive since in an OsmoTRX based BTS model those frames
will be generated by OsmoTRX and do not appear on the PCUIF interface.
When the BTS model is left unspecified, then the PCU will generate those
DUMMY messages and irretate the testsuite.

Related: OS#6191
Change-Id: I6b516eece6acc96ebde9759bcb609197a245dd84