GTPv2_Emulation: Improve routing messages to client
Before this patch, for instance, registering an IMSI would not
necessarily mean the messaged ended up in the client registering it, due
to how the code paths were being handled.

Ideally the code should create a CLIENT candidate list based on all
registered matches, and then remove duplicates and finally send message
to each CLIENT in the resulting list. That's really not needed as of now
though.

Change-Id: If253a2b8991f0caf685058238df9a7c997983e27