SIP_Emulation: Remove duplicated map()
The map() is already done before the component is started, f_init_sip().
It's better doing it there because then it's up to the creator of the
SIP_Emulation component to establish the name of the system port, hence
allowing multiple instances which can be configured differently (eg.
different bind ip+port).

Change-Id: I5632c94d03a9dce0d04d94766ce72e6700b6116b