SIP_Emulation: Rename component id
Remove the dash character, since it makes it impossible to reference the
component name under TESTPORT_PARAMETERS in .default/.cfg files.

While at it, actually, allow the test to provide a full id instead of
always appending the suffix to it. This provides more freedom on the
testsuite to provide a fitting component name.

Change-Id: Iecefe7d98a5842872f1efc55e013f672186ef1a8