SGP32Definitions_Templates: improve ts_initiateAuthenticationRequestEsipa
The template ts_initiateAuthenticationRequestEsipa has a field for smdpAddress and
euiccInfo1. Those fields are optional fields, but they still play a central role in
the protocol, so that they are effectively mandatory. Let's make the fields available
to the template user and populate them with meaningful default values.

Related: SYS#6824
Change-Id: I97aa039810ad9fffea4226254fa675fd24647de1