asterisk: Allow passing full templates to {tr,ts}_SIP_msgh_std
It may have been a great idea at first to have this SipAddr record
unifying several records, and it may be so under certain circumstances,
but clearly it's not the case when passing them to generic tr_* templates
(like the _std_ one in SIP) which then have to access specific fields of
those records. This will crash if for instance a (tr_...(), tr_...())
compound is passed.

This is a next step towards getting rid of SipAddress in lower level
templates, which the SIP specific enc/dec doesn't understand and creates
several kinds of troubles like the one mentioned above.

Change-Id: I3a85701822d7b76ec078b1a7c615e5ef50f36902