SGP32Definitions_Templates: use signature dummys with correct length
The dummy signature values we are currently using do not have the correct length.
The ECDSA signatures used in SGP.32 have a well specified lengt of 64 bytes and
they consist of two 32 bit integer nummbers called "R" and "S", so let's use
signature dummys with the correct length and recognizeable values.

Related: SYS#6824
Change-Id: Ieea99bab5c926420ce864d7c5067e55b2861b045