testenv: podman.is_running() -> testenv.args.podman
Use testenv.args.podman instead of testenv.podman.is_running() in all
places except for testenv.podman.stop().

- testenv.args.podman is always True when --podman is used.

- testenv.podman.is_running() is only True while the container is
  currently running.

Most of the time the behavior is the same. But without this patch, when
the container crashes, commands would unexpectedly run outside of the
container (and then fail on jenkins due to missing programs).

Related: OS#6494
Change-Id: Iea634f5e97d14f1f7e777f4609b9593974964d23