testenv: rename podman dir to virt
Prepare to add QEMU as second virtualization next to podman, which can
be used to test e.g. the GTP kernel driver.

QEMU will use many of the same paths that are currently used for podman,
so rename "podman" to "virt" for virtualization. I chose the name,
because podman is paravirtualization and QEMU is full virtualization.

Change-Id: I22ecc922297d598d09eebc8b36dd2e1a0726463c