mme: expect authentication during 2g->4G TAU procedure
If TAU happens without an existing security context, network for sure wants
to authenticate the UE. This is the scenario being tested here.

If there was a security context (eg. from a previous 4G Attach), then
it's up to the network to decide (or be capable) to implicitly
authenticate the UE.

Change-Id: I5ebcaf23a643b6cb97534328737257d26d3981fe