S1AP_Emulation: Fix building security headers for Integrity and Ciphering
The fact that the network instructs the UE to use EEA0 encryption (null
algo) doesn't mean the UE should mark the NAS frames in the security
header as non-ciphered. They need to be marked as IP+ciphered in the
outer header, and as plain in the inner one.

Now TTCN3 behaves like a UE recorded while executing a similar test
case. Moreover, wireshark is now happy decoding the messages.

Change-Id: Ifd4fe83e7b5aefdaafe7ce1c8b1ec6a67c65819c