ggsn: README: document how to run with testenv
It is possible to run the GGSN testsuite in a lot of ways (as it was
ported from docker-playground, which had the same variations but with a
less consistent way of running them).

Document how it is typically run for development / in jenkins. This
should make it easier for users, in addition to testenv already telling
which configs are available if trying to run the ggsn testsuite without
the -c argument, and to the general help output in "./testenv.py run -h".

$ ./testenv.py run ggsn
[testenv] Using testsuite ggsn_tests (via alias ggsn)
[testenv] Found multiple testenv.cfg files:
[testenv]  * testenv_open5gs.cfg
[testenv]  * testenv_osmo_ggsn_all.cfg
[testenv]  * testenv_osmo_ggsn_v4_only.cfg
[testenv]  * testenv_osmo_ggsn_v4v6_only.cfg
[testenv]  * testenv_osmo_ggsn_v6_only.cfg
[testenv] Select a specific config (e.g. '-c open5gs') or all ('-c all')

Change-Id: I49421151b199fa6536da784b6c412b03f3fdbd05