testenv: improve output for missing -c argument
When a testsuite has multiple testenv.cfg files, the user needs to
explicitly choose a config, or "-c all" for all configs. Improve the
help output to directly print the arguments that need to be passed,
instead of printing the config file names. Mention that wildcards can be
used too.

Old:
  [testenv] Found multiple testenv.cfg files:
  [testenv]  * testenv_generic.cfg
  [testenv]  * testenv_sccplite.cfg
  [testenv]  * testenv_vamos.cfg
  [testenv] Select a specific config (e.g. '-c generic') or all ('-c all')

New:
  [testenv] Found multiple testenv.cfg files, use one of:
  [testenv]  -c generic
  [testenv]  -c sccplite
  [testenv]  -c vamos
  [testenv] You can also select all of them (-c all) or use the * character as wildcard.

Related: OS#6494
Change-Id: I3f273caff702b33d3d74a9e5c8d77b22f27d7cfc