testenv: podman-main: set watchdog timer to 60s
Increase the timer from 10s to 60s, as with 10s I see jobs failing with:

  ERROR: /tmp/watchdog was not created, exiting

In theory 10s should already be enough, my guess is that if a jenkins
node is currently under a lot of load then the feed command may take
several seconds and so we hit the previous timer. Even if this is not
the cause, I think it is good to rule it out.

Exiting after 60s if the jenkins job was (manually / with connection
loss) aborted is still relatively quick.

Related: OS#6607
Change-Id: Ide7a9d5bc395106cdaa47cec3aa746a978674c9e