testenv: show testsuite results at the end
Show the testsuite results from junit-*.log not after each testenv*.cfg
file is through, but for all of them at the end. This way the results
are in one place when running with multiple configs, we don't need to
scroll to the middle of the huge log to find out what tests passed with
a previous config.

Adjust the podman container stop and restart logic, so the last
container is still running until we use it for showing the results.

Change-Id: Ie4b70724a33932cd98ba0972db37bf72dc8affb8