testenv: build virtphy from src with --binary-repo
Add logic to build virtphy from source when running with --binary-repo.
This extra code path is needed because we currently don't have virtphy
packaged (like trxcon and sccp_demo_user), and we need to build the
libosmocore binary package instead of building completely from source as
we would do it with osmo-dev.

Use ".split(" ", 1)[0]" on the program= value to only look at its first
word, so we can later on use it in testenv.cfg file as follows:

  [virtphy]
  program=virtphy -s /tmp/osmocom_l2

Change-Id: I37bac8509b2601286e4feab099782f82c8338dca