testenv: make copy= more intuitive
Currently copy= in testenv.cfg creates copies of files under the full
source file name under the target directory: copy=dir/file.cfg creates
dir/file.cfg (like "cp -a --parents"). This is not very intuitive,
change it to create a "file.cfg" instead without subdirectory. With
this change, it behaves the same as "cp -a".

Suggested-by: Pau Espin Pedrol <pespin@sysmocom.de>
Change-Id: Id9e4ea066d0a0e72ccd83e4f6f887f382bf61e46