RAN_Adapter: Create components alive
HNBGW_Tests.TC_hnb_disconnected_timeout needs modification since it
expects the component to drop the underlaying conn towards the IUT when
the component is stopped. This is not longer the case when the component
is created as "alive". In order to make sure its resources are
destroyed, one needs to kill it.

Change-Id: Ice4d93b42552f1c5d621582b1fc3a38b4e1ae837