HTTP_Adapter: allow requests to different hosts
The host that is requested via the HTTP_Adapter is configured once on
initialization. This is fine if the test scenario only has exactly one
destination to query. For multiple destinations, this model does not
work. Let's add an http_pars parameter to the request functions, so
that the user can direct the requests to different hosts dynamically.

Related: SYS#7339
Change-Id: Ic87674f2381b3c6a1be6d0ce28a9e5403cda201b